requirements.txt
requests
cloudflare-ddns
loguru
python-crontab
python-telegram-bot